requirements.txt
Flask
gunicorn
markdown2
supabase
python-dotenv
